refactor: simplify jq filter initialization and remove unused libpath
