+++ go.mod
@@ -5,7 +5,6 @@ go 1.23.1
 require (
 	github.com/deckhouse/deckhouse/pkg/log v0.0.0-20241205040953-7b376bae249c
 	github.com/flant/kube-client v1.2.2
-	github.com/flant/libjq-go v1.6.3-0.20201126171326-c46a40ff22ee // branch: master
 	github.com/go-chi/chi/v5 v5.2.1
 	github.com/go-openapi/spec v0.19.8
 	github.com/go-openapi/strfmt v0.19.5
@@ -36,6 +35,7 @@ replace github.com/go-openapi/validate => github.com/flant/go-openapi-validate v
 require (
 	github.com/deckhouse/module-sdk v0.2.0
 	github.com/gojuno/minimock/v3 v3.4.5
+	github.com/itchyny/gojq v0.12.17
 )
 
 require (
@@ -71,6 +71,7 @@ require (
 	github.com/gregjones/httpcache v0.0.0-20180305231024-9cad4c3443a7 // indirect
 	github.com/hashicorp/errwrap v1.1.0 // indirect
 	github.com/imdario/mergo v0.3.16 // indirect
+	github.com/itchyny/timefmt-go v0.1.6 // indirect
 	github.com/jonboulle/clockwork v0.4.0 // indirect
 	github.com/josharian/intern v1.0.0 // indirect
 	github.com/json-iterator/go v1.1.12 // indirect
