+++ pkg/app/app.go
@@ -134,7 +134,6 @@ func DefineStartCommandFlags(kpApp *kingpin.Application, cmd *kingpin.CmdClause)
 	DefineKubeClientFlags(cmd)
 	DefineValidatingWebhookFlags(cmd)
 	DefineConversionWebhookFlags(cmd)
-	DefineJqFlags(cmd)
 	DefineLoggingFlags(cmd)
 	DefineDebugFlags(kpApp, cmd)
 }
