+++ pkg/kube_events_manager/filter_test.go
@@ -13,7 +13,7 @@ import (
 func TestApplyFilter(t *testing.T) {
 	t.Run("filter func with error", func(t *testing.T) {
 		uns := &unstructured.Unstructured{Object: map[string]interface{}{"foo": "bar"}}
-		filter := jq.NewFilter("")
+		filter := jq.NewFilter()
 		_, err := applyFilter("", filter, filterFuncWithError, uns)
 		assert.EqualError(t, err, "filterFn (github.com/flant/shell-operator/pkg/kube_events_manager.filterFuncWithError) contains an error: invalid character 'a' looking for beginning of value")
 	})
